requirements.txt
absl-py==0.4.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
chrome-gnome-shell==0.0.0
cloudpickle==0.5.6
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dask==0.19.1
decorator==4.3.0
defer==1.0.6
distro-info==0.18
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
httplib2==0.9.2
idna==2.6
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
networkx==2.1
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
pandas==0.23.4
pbr==4.2.0
pexpect==4.2.1
Pillow==5.1.0
pkg-resources==0.0.0
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.2
python-dateutil==2.7.3
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.0.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
SecretStorage==2.3.1
selectivesearch==0.4
simplejson==3.13.2
six==1.11.0
sklearn==0.0
system-service==0.3
systemd-python==234
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
testresources==2.0.1
toolz==0.9.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==0.14.1
xkit==0.0.0
zope.interface==4.3.2
